refactor(agentic-ai): move ChatModelProviderSupport out of the langchain4j package
CONNECT_TIMEOUT and deriveTimeoutSetting are shared by every chat
model factory, native and LangChain4j-routed alike, and never
reference dev.langchain4j.* — but living under
chatmodel.provider.langchain4j.factory made the native Bedrock
factory import from the LangChain4j package, contradicting the
framework-agnostic-core invariant (only chatmodel/provider/langchain4j/**
may import dev.langchain4j.*). Moved to the neutral parent
chatmodel.provider package.

Also fills in the chatmodel.provider.bedrock.configuration
package-info.java gap the generator script found while regenerating.
